Cleanup: move key rotation routines to where they are used
