+++ subcommands/keys/tuf_rotate_offline_key.go
@@ -3,14 +3,9 @@ package keys
 import (
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/spf13/cobra"
 
-	canonical "github.com/docker/go/canonical/json"
-	tuf "github.com/theupdateframework/notary/tuf/data"
-
-	"github.com/foundriesio/fioctl/client"
 	"github.com/foundriesio/fioctl/subcommands"
 )
 
@@ -166,82 +161,4 @@ func doRotateOfflineKey(cmd *cobra.Command, args []string) {
 	fmt.Println("= Applying staged TUF root changes")
 	tufUpdatesCmd.SetArgs([]string{"apply"})
 	subcommands.DieNotNil(tufUpdatesCmd.Execute())
-}
-
-func swapRootKey(
-	root *client.AtsTufRoot, creds OfflineCreds, keyType TufKeyType,
-) (*TufSigner, OfflineCreds) {
-	kp := genTufKeyPair(keyType)
-	root.Signed.Keys[kp.signer.Id] = kp.atsPub
-	root.Signed.Expires = time.Now().AddDate(1, 0, 0).UTC().Round(time.Second) // 1 year validity
-	root.Signed.Roles["root"].KeyIDs = []string{kp.signer.Id}
-
-	base := "tufrepo/keys/fioctl-root-" + kp.signer.Id
-	creds[base+".pub"] = kp.atsPubBytes
-	creds[base+".sec"] = kp.atsPrivBytes
-	return &kp.signer, creds
-}
-
-func findOnlineTargetsId(factory string, root client.AtsTufRoot) (string, error) {
-	onlinePub, err := api.TufTargetsOnlineKey(factory)
-	subcommands.DieNotNil(err)
-	for _, keyid := range root.Signed.Roles["targets"].KeyIDs {
-		pub := root.Signed.Keys[keyid].KeyValue.Public
-		if pub == onlinePub.KeyValue.Public {
-			return keyid, nil
-		}
-	}
-	return "", errors.New("Unable to find online target key for factory")
-}
-
-func replaceOfflineTargetsKey(
-	root *client.AtsTufRoot, onlineTargetsId string, creds OfflineCreds, keyType TufKeyType,
-) (*TufSigner, OfflineCreds) {
-	kp := genTufKeyPair(keyType)
-	root.Signed.Keys[kp.signer.Id] = kp.atsPub
-	root.Signed.Roles["targets"].KeyIDs = []string{onlineTargetsId, kp.signer.Id}
-	root.Signed.Roles["targets"].Threshold = 1
-
-	base := "tufrepo/keys/fioctl-targets-" + kp.signer.Id
-	creds[base+".pub"] = kp.atsPubBytes
-	creds[base+".sec"] = kp.atsPrivBytes
-	return &kp.signer, creds
-}
-
-func resignProdTargets(
-	factory string, root *client.AtsTufRoot, onlineTargetsId string, creds OfflineCreds,
-) (map[string][]tuf.Signature, error) {
-	targetsMap, err := api.ProdTargetsList(factory, false)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to fetch production targets: %w", err)
-	} else if targetsMap == nil {
-		return nil, nil
-	}
-
-	var signers []TufSigner
-	for _, kid := range root.Signed.Roles["targets"].KeyIDs {
-		if kid == onlineTargetsId {
-			continue
-		}
-		pub := root.Signed.Keys[kid].KeyValue.Public
-		signer, err := FindTufSigner(kid, pub, creds)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to find private key for %s: %w", kid, err)
-		}
-		signers = append(signers, *signer)
-	}
-
-	signatureMap := make(map[string][]tuf.Signature)
-	for tag, targets := range targetsMap {
-		bytes, err := canonical.MarshalCanonical(targets.Signed)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to marshal targets for tag %s: %w", tag, err)
-		}
-		signatures, err := SignTufMeta(bytes, signers...)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to re-sign targets for tag %s: %w", tag, err)
-		}
-		signatureMap[tag] = signatures
-	}
-	return signatureMap, nil
 }
